requirements.txt
Pillow==9.3.0
pyperclip==1.8.2
pyttsx3==2.90
